% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{access_signature}
\alias{access_signature}
\title{Generate an api access signature}
\usage{
access_signature(timestamp, request_path, request_body = "",
  method = c("get", "post", "put", "delete"))
}
\arguments{
\item{timestamp}{The CB-ACCESS-TIMESTAMP header MUST be number of seconds since Unix Epoch.}

\item{request_path}{path requesting}

\item{request_body}{The body is the request body string or omitted if there is no request body (typically for GET requests).}

\item{method}{The method should be UPPER CASE.}
}
\description{
The CB-ACCESS-SIGN header is generated by creating a sha256 HMAC using the
secret key on the prehash string timestamp + method + requestPath + body
(where + represents string concatenation). The timestamp value is the same as
 the CB-ACCESS-TIMESTAMP header.
}
\keyword{internal}
