% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{assign_tokens}
\alias{assign_tokens}
\title{Assign working tokens}
\usage{
assign_tokens()
}
\value{
creates working \code{environment} that reads/writes binary tokens and assigns variable in working environment
}
\description{
Assign working tokens
}
\examples{
\dontrun{
  assign_tokens()
}
}
