% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_jwt.R
\name{build_jwt}
\alias{build_jwt}
\title{Gets Bearer Token}
\usage{
build_jwt(key_var, secret_var, method, endpoint)
}
\arguments{
\item{key_var}{=  your personal API key}

\item{secret_var}{= your personal secret token}

\item{method}{= GET, POST, etc.}

\item{endpoint}{= endpoint to use}
}
\value{
returns JWT token to make API requests
}
\description{
Gets Bearer Token
}
