% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{capture_request2}
\alias{capture_request2}
\title{Capture requests to tests directory}
\usage{
capture_request2(expr)
}
\arguments{
\item{expr}{R expresssion to execute}
}
\description{
Helper to build up mock fixtures for testing
}
\examples{
\dontrun{
capture_requests2(get_accounts())
}
}
