% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_cancel_futures_sweep}
\alias{cb_cancel_futures_sweep}
\title{Futures: Cancel Sweep}
\usage{
cb_cancel_futures_sweep()
}
\value{
A \code{data.frame} detailing the pending sweep of funds from FCM wallet to USD Spot wallet
}
\description{
Futures: Cancel Sweep
}
\examples{
\dontrun{
  cb_cancel_futures_sweep()
}
}
