% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_cancel_order}
\alias{cb_cancel_order}
\title{Spot: Cancel Order}
\usage{
cb_cancel_order(order_ids)
}
\arguments{
\item{order_ids}{= (string) enter the order id that you wish to cancel}
}
\value{
returns order details as a \code{data.frame} for cancelled orders
}
\description{
Spot: Cancel Order
}
\examples{
\dontrun{
  cb_cancel_order(order_ids='ASDGF123-SDVSA123-SAEF123')
}
}
