% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_close_order}
\alias{cb_close_order}
\title{Futures: Cancel Order}
\usage{
cb_close_order(client_order_id = cb_get_order_id(), product_id, size = NULL)
}
\arguments{
\item{client_order_id}{= defaults to random id via \emph{cb_get_order_id()}}

\item{product_id}{= futures contract to close}

\item{size}{= number of contracts to close, defaults to closing all available}
}
\value{
Cancel response \code{data.frame} status for a futures order
}
\description{
Futures: Cancel Order
}
\examples{
\dontrun{
  cb_close_order(product_id = "BIT-28JUL23-CDE")
}
}
