% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_get_current_margin_window}
\alias{cb_get_current_margin_window}
\title{Futures: Get Margin Window}
\usage{
cb_get_current_margin_window(
  margin_profile_type = "MARGIN_PROFILE_TYPE_RETAIL_REGULAR"
)
}
\arguments{
\item{margin_profile_type}{= The margin profile type for your account: \emph{MARGIN_PROFILE_TYPE_UNSPECIFIED}}
}
\value{
Get the futures current margin window as a \code{data.frame}
}
\description{
Futures: Get Margin Window
}
\examples{
\dontrun{
  cb_get_current_margin_window()
}
}
