% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_get_fees}
\alias{cb_get_fees}
\title{Account Fees}
\usage{
cb_get_fees(
  product_type = "UNKNOWN_PRODUCT_TYPE",
  contract_expiry_type = "UNKNOWN_CONTRACT_EXPIRY_TYPE",
  product_venue = "UNKNOWN_VENUE_TYPE"
)
}
\arguments{
\item{product_type}{= Only returns the orders matching this product type.
By default, returns all product types.
'UNKNOWN_PRODUCT_TYPE', 'SPOT','FUTURE'}

\item{contract_expiry_type}{= Only returns the orders matching this contract expiry type.
Only applicable if product_type is set to FUTURE.
'UNKNOWN_CONTRACT_EXPIRY_TYPE','EXPIRING', 'PERPETUAL'}

\item{product_venue}{= Venue for product
'UNKNOWN_VENUE_TYPE','CBE','FCM','INTX'}
}
\value{
A \code{data.frame} with a summary of transactions with fee tiers, total volume, and fees.
}
\description{
Account Fees
}
\examples{
\dontrun{
  cb_get_fees()
}
}
