% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_get_futures_balance}
\alias{cb_get_futures_balance}
\title{Futures: Get Balance}
\usage{
cb_get_futures_balance()
}
\value{
Get Futures Balance Summary as a \code{data.frame}
}
\description{
Futures: Get Balance
}
\examples{
\dontrun{
  cb_get_futures_balance()
}
}
