% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_get_futures_position}
\alias{cb_get_futures_position}
\title{Futures: Get Position}
\usage{
cb_get_futures_position(product_id)
}
\arguments{
\item{product_id}{= The ticker symbol (e.g. 'BIT-28JUL25-CDE')}
}
\value{
Get Futures Position as a \code{data.frame}

Get positions for a specific CFM product
}
\description{
Futures: Get Position
}
\examples{
\dontrun{
  cb_get_futures_position(product_id = 'BIT-28JUL25-CDE')
}
}
