% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_get_intraday_margin}
\alias{cb_get_intraday_margin}
\title{Futures: Get Intraday Margin}
\usage{
cb_get_intraday_margin()
}
\value{
Get intraday margin Setting as a \code{data.frame}
}
\description{
Futures: Get Intraday Margin
}
\examples{
\dontrun{
  cb_get_intraday_margin()
}
}
