% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_get_order_id}
\alias{cb_get_order_id}
\title{Order ID}
\usage{
cb_get_order_id()
}
\value{
An auto generated \code{character} string to use for placing orders
}
\description{
Order ID
}
\examples{
\dontrun{
  cb_get_order_id()
}
}
