% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_getAccount}
\alias{cb_getAccount}
\title{Get Accounts}
\usage{
cb_getAccount(acct_uuid)
}
\arguments{
\item{acct_uuid}{= The account's UUID.}
}
\value{
Get a \code{data.frame} of information about an account, given an account UUID.
}
\description{
Get Accounts
}
\examples{
\dontrun{
  cb_getAccount(acct_uuid = 'f412dr89-01d0-576d-g457-ea0b52a13716')
}
}
