% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_getAccounts}
\alias{cb_getAccounts}
\title{List Accounts}
\usage{
cb_getAccounts(lmt)
}
\arguments{
\item{lmt}{= The number of accounts to display per page. By default, displays 49 (max 250)}
}
\value{
Get a \code{data.frame} of authenticated Advanced Trade accounts for the current user.
}
\description{
List Accounts
}
\examples{
\dontrun{
  cb_getAccounts(lmt = 100)
}
}
