% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_getCryptoList}
\alias{cb_getCryptoList}
\title{Get Crypto List}
\usage{
cb_getCryptoList()
}
\value{
Get a \code{data.frame} with all crypto currency pairs
}
\description{
Get Crypto List
}
\examples{
\dontrun{
  cb_getCryptoList()
}
}
