% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_getOrder}
\alias{cb_getOrder}
\title{Get Order}
\usage{
cb_getOrder(id)
}
\arguments{
\item{id}{= The ID of the order}
}
\value{
Get a detailed \code{data.frame} for the order requested.
}
\description{
Get Order
}
\examples{
\dontrun{
  cb_getOrder(id='1234')
}
}
