% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_list_futures_positions}
\alias{cb_list_futures_positions}
\title{Futures: List All Positions}
\usage{
cb_list_futures_positions()
}
\value{
Get a list of positions in CFM products as a \code{data.frame}
}
\description{
Futures: List All Positions
}
\examples{
\dontrun{
  cb_list_futures_positions()
}
}
