% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_list_futures_sweeps}
\alias{cb_list_futures_sweeps}
\title{Futures: List Sweeps}
\usage{
cb_list_futures_sweeps()
}
\value{
Gets \code{data.frame} for pending and processing sweeps of funds from FCM wallet to USD Spot wallet
}
\description{
Futures: List Sweeps
}
\examples{
\dontrun{
  cb_list_futures_sweeps()
}
}
