% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_quote}
\alias{cb_quote}
\title{Crypto Pair Bid/Ask Quotes}
\usage{
cb_quote(ids)
}
\arguments{
\item{ids}{= vector of product id(s) Example: "BTC-USD" OR c("BTC-USD","ETH-USD")}
}
\value{
Get a \code{data.frame} for the best bid/ask for all products.
}
\description{
Crypto Pair Bid/Ask Quotes
}
\examples{
\dontrun{
  cb_quote(ids=c("BTC-USD","ETH-USD"))
}
}
