% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_schedule_futures_sweeps}
\alias{cb_schedule_futures_sweeps}
\title{Futures: Schedule Sweeps}
\usage{
cb_schedule_futures_sweeps(usd_amount)
}
\arguments{
\item{usd_amount}{= The amount of USD to be swept. By default, sweeps all available excess funds.}
}
\value{
Gets \code{data.frame} for scheduling a sweep of funds from FCM wallet to USD Spot wallet
}
\description{
Futures: Schedule Sweeps
}
\examples{
\dontrun{
  cb_schedule_futures_sweeps(usd_amount = 100.00)
}
}
