% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCoinBase.R
\name{cb_set_intraday_margin}
\alias{cb_set_intraday_margin}
\title{Futures: Set Intraday Margin}
\usage{
cb_set_intraday_margin(setting)
}
\arguments{
\item{setting}{= The amount of USD to be swept. By default, sweeps all available excess funds.}
}
\value{
Gets \code{data.frame} with details for setting intraday margin
}
\description{
Futures: Set Intraday Margin
}
\examples{
\dontrun{
  cb_set_intraday_margin(setting = 100.00)
}
}
