% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-accounts.r
\name{create_account}
\alias{create_account}
\title{Create account}
\usage{
create_account(body)
}
\arguments{
\item{body}{a named list of parameters}
}
\description{
Creates a new account for user.
}
\examples{
\dontrun{
body <- list(name = "New Wallet")
resp <- create_account(body)
}
}
\references{
\url{https://developers.coinbase.com/api/v2#create-account}
}
\seealso{
Other wallet-accounts: \code{\link{delete_account}},
  \code{\link{get_accounts}}, \code{\link{get_account}},
  \code{\link{get_address_transactions}},
  \code{\link{get_address}},
  \code{\link{set_primary_account}},
  \code{\link{update_account}}
}
\keyword{internal}
