% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-accounts.r
\name{delete_account}
\alias{delete_account}
\title{Delete account}
\usage{
delete_account(account_id)
}
\arguments{
\item{account_id}{Account Id}
}
\description{
Removes user’s account. In order to remove an account it can’t be:
\itemize{
\item Primary account
\item Account with non-zero balance
\item Fiat account
\item Vault with a pending withdrawal
}
}
\examples{
\dontrun{
resp <- delete_account("123456")
}
}
\references{
\url{https://developers.coinbase.com/api/v2#delete-account}
}
\seealso{
Other wallet-accounts: \code{\link{create_account}},
  \code{\link{get_accounts}}, \code{\link{get_account}},
  \code{\link{get_address_transactions}},
  \code{\link{get_address}},
  \code{\link{set_primary_account}},
  \code{\link{update_account}}
}
\keyword{internal}
