% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-accounts.r
\name{get_account}
\alias{get_account}
\title{Show an account}
\usage{
get_account(account_id)
}
\arguments{
\item{account_id}{Account Id}
}
\description{
Show current user’s account. To access the primary account for a given
currency, a currency string (BTC or ETH) can be used instead of the account
id in the URL.
}
\references{
\url{https://developers.coinbase.com/api/v2#show-an-account}
}
\seealso{
Other wallet-accounts: \code{\link{create_account}},
  \code{\link{delete_account}}, \code{\link{get_accounts}},
  \code{\link{get_address_transactions}},
  \code{\link{get_address}},
  \code{\link{set_primary_account}},
  \code{\link{update_account}}
}
