% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-accounts.r
\name{get_accounts}
\alias{get_accounts}
\title{List accounts}
\usage{
get_accounts()
}
\description{
Lists current user’s accounts to which the authentication method has access to.
}
\references{
\url{https://developers.coinbase.com/api/v2#list-accounts}
}
\seealso{
Other wallet-accounts: \code{\link{create_account}},
  \code{\link{delete_account}}, \code{\link{get_account}},
  \code{\link{set_primary_account}},
  \code{\link{update_account}}
}
