% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-addresses.r
\name{get_address}
\alias{get_address}
\title{Show addresss}
\usage{
get_address(account_id, address_id)
}
\arguments{
\item{account_id}{Account Id}

\item{address_id}{Address Id}
}
\description{
Show an individual address for an account. A regular bitcoin, litecoin or
ethereum address can be used in place of address_id but the address has to
be associated to the correct account.
}
\references{
\url{https://developers.coinbase.com/api/v2#show-addresss}
}
\seealso{
Other wallet-accounts: \code{\link{create_account}},
  \code{\link{delete_account}}, \code{\link{get_accounts}},
  \code{\link{get_account}},
  \code{\link{get_address_transactions}},
  \code{\link{set_primary_account}},
  \code{\link{update_account}}
}
