% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-addresses.r
\name{get_address_transactions}
\alias{get_address_transactions}
\title{List address’s transactions}
\usage{
get_address_transactions(account_id, address_id)
}
\arguments{
\item{account_id}{Account Id}

\item{address_id}{Address Id}
}
\description{
List transactions that have been sent to a specific address.
A regular bitcoin, litecoin or ethereum address can be used in place of
address_id but the address has to be associated to the correct account.
}
\references{
\url{https://developers.coinbase.com/api/v2#list-address39s-transactions}
}
\seealso{
Other wallet-addresses: \code{\link{create_address}},
  \code{\link{get_addresses}}, \code{\link{get_address}}
}
