% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-addresses.r
\name{get_addresses}
\alias{get_addresses}
\title{List addresses}
\usage{
get_addresses(account_id)
}
\arguments{
\item{account_id}{Account Id}
}
\description{
Lists addresses for an account.
}
\references{
\url{https://developers.coinbase.com/api/v2#list-addresses}
}
\seealso{
Other wallet-addresses: \code{\link{create_address}},
  \code{\link{get_address_transactions}},
  \code{\link{get_address}}
}
