% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-users.r
\name{get_auth_info}
\alias{get_auth_info}
\title{Show authorization information}
\usage{
get_auth_info()
}
\description{
Get current user’s authorization information including granted scopes and
send limits when using OAuth2 authentication.
}
\references{
\url{https://developers.coinbase.com/api/v2#show-authorization-information}
}
\seealso{
Other wallet-users: \code{\link{get_current_user}},
  \code{\link{get_user}}, \code{\link{update_current_user}}
}
