% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-buys.r
\name{get_buy}
\alias{get_buy}
\title{Show a buy}
\usage{
get_buy(account_id, buy_id)
}
\arguments{
\item{account_id}{Account Id}

\item{buy_id}{Buy Id}
}
\description{
Show an individual buy.
}
\references{
\url{https://developers.coinbase.com/api/v2#show-a-buy}
}
\seealso{
Other wallet-buys: \code{\link{get_buys}}
}
