% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-buys.r
\name{get_buys}
\alias{get_buys}
\title{List buys}
\usage{
get_buys(account_id)
}
\arguments{
\item{account_id}{Account Id}
}
\description{
Lists buys for an account.
}
\references{
\url{https://developers.coinbase.com/api/v2#list-buys}
}
\seealso{
Other wallet-buys: \code{\link{get_buy}}
}
