% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-users.r
\name{get_current_user}
\alias{get_current_user}
\title{Show current user}
\usage{
get_current_user()
}
\description{
Get current user’s public information. To get user’s email or private
information, use permissions wallet:user:email and wallet:user:read. If
current request has a wallet:transactions:send scope, then the response
will contain a boolean sends_disabled field that indicates if the user’s
send functionality has been disabled.
}
\references{
\url{https://developers.coinbase.com/api/v2#show-current-user}
}
\seealso{
Other wallet-users: \code{\link{get_auth_info}},
  \code{\link{get_user}}, \code{\link{update_current_user}}
}
