% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-deposits.r
\name{get_deposit}
\alias{get_deposit}
\title{Show a deposit}
\usage{
get_deposit(account_id, deposit_id)
}
\arguments{
\item{account_id}{Account Id}

\item{deposit_id}{Deposit Id}
}
\description{
Show an individual deposit.
}
\references{
\url{https://developers.coinbase.com/api/v2#show-a-deposit}
}
\seealso{
Other wallet-deposits: \code{\link{get_deposits}}
}
