% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-deposits.r
\name{get_deposits}
\alias{get_deposits}
\title{List deposits}
\usage{
get_deposits(account_id)
}
\arguments{
\item{account_id}{Account Id}
}
\description{
Lists deposits for an account.
}
\references{
\url{https://developers.coinbase.com/api/v2#list-deposits}
}
\seealso{
Other wallet-deposits: \code{\link{get_deposit}}
}
