% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-payment-methods.r
\name{get_payment_method}
\alias{get_payment_method}
\title{Show a payment method}
\usage{
get_payment_method(payment_method_id)
}
\arguments{
\item{payment_method_id}{Payment Method Id}
}
\description{
Show a current user's payment method.
}
\references{
\url{https://developers.coinbase.com/api/v2#show-a-payment-method}
}
\seealso{
Other wallet-payment-methods: \code{\link{get_payment_methods}}
}
