% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-payment-methods.r
\name{get_payment_methods}
\alias{get_payment_methods}
\title{List payment methods}
\usage{
get_payment_methods()
}
\description{
Lists current user's payment methods.
}
\references{
\url{https://developers.coinbase.com/api/v2#list-payment-methods}
}
\seealso{
Other wallet-payment-methods: \code{\link{get_payment_method}}
}
