% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-sells.r
\name{get_sell}
\alias{get_sell}
\title{Show a sell}
\usage{
get_sell(account_id, sell_id)
}
\arguments{
\item{account_id}{Account Id}

\item{sell_id}{Sell Id}
}
\description{
Show an individual sell.
}
\references{
\url{https://developers.coinbase.com/api/v2#show-a-sell}
}
\seealso{
Other wallet-sells: \code{\link{get_sells}}
}
