% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-sells.r
\name{get_sells}
\alias{get_sells}
\title{List sells}
\usage{
get_sells(account_id)
}
\arguments{
\item{account_id}{Account Id}
}
\description{
Lists sells for an account.
}
\references{
\url{https://developers.coinbase.com/api/v2#list-sells}
}
\seealso{
Other wallet-sells: \code{\link{get_sell}}
}
