% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-transactions.r
\name{get_transaction}
\alias{get_transaction}
\title{Show a transaction}
\usage{
get_transaction(account_id, transaction_id)
}
\arguments{
\item{account_id}{Account Id}

\item{transaction_id}{Transaction Id}
}
\description{
Show an individual transaction for an account.
}
\references{
\url{https://developers.coinbase.com/api/v2#show-a-transaction}
}
\seealso{
Other wallet-transactions: \code{\link{get_transactions}}
}
