% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-transactions.r
\name{get_transactions}
\alias{get_transactions}
\title{List transactions}
\usage{
get_transactions(account_id)
}
\arguments{
\item{account_id}{Account Id}
}
\description{
Lists account's transactions.
}
\references{
\url{https://developers.coinbase.com/api/v2#list-transactions}
}
\seealso{
Other wallet-transactions: \code{\link{get_transaction}}
}
