% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-users.r
\name{get_user}
\alias{get_user}
\title{Show a user}
\usage{
get_user(user_id)
}
\arguments{
\item{user_id}{User Id}
}
\description{
Get any user’s public information with their ID.
}
\references{
\url{https://developers.coinbase.com/api/v2#show-a-user}
}
\seealso{
Other wallet-users: \code{\link{get_auth_info}},
  \code{\link{get_current_user}},
  \code{\link{update_current_user}}
}
