% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-withdrawals.r
\name{get_withdrawal}
\alias{get_withdrawal}
\title{Show a withdrawal}
\usage{
get_withdrawal(account_id, withdrawal_id)
}
\arguments{
\item{account_id}{Account Id}

\item{withdrawal_id}{Withdrawal Id}
}
\description{
Show an individual withdrawal.
}
\references{
\url{https://developers.coinbase.com/api/v2#show-a-withdrawal}
}
\seealso{
Other wallet-withdrawals: \code{\link{get_withdrawals}}
}
