% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-withdrawals.r
\name{get_withdrawals}
\alias{get_withdrawals}
\title{List withdrawals}
\usage{
get_withdrawals(account_id)
}
\arguments{
\item{account_id}{Account Id}
}
\description{
Lists withdrawals for an account.
}
\references{
\url{https://developers.coinbase.com/api/v2#list-withdrawals}
}
\seealso{
Other wallet-withdrawals: \code{\link{get_withdrawal}}
}
