% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-accounts.r
\name{set_primary_account}
\alias{set_primary_account}
\title{Set account as primary}
\usage{
set_primary_account(account_id)
}
\arguments{
\item{account_id}{Account Id}
}
\description{
Promote an account as primary account.
}
\examples{
\dontrun{
resp <- set_primary_account("123456")
}
}
\references{
\url{https://developers.coinbase.com/api/v2#set-account-as-primary}
}
\seealso{
Other wallet-accounts: \code{\link{create_account}},
  \code{\link{delete_account}}, \code{\link{get_accounts}},
  \code{\link{get_account}},
  \code{\link{get_address_transactions}},
  \code{\link{get_address}}, \code{\link{update_account}}
}
\keyword{internal}
