% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-accounts.r
\name{update_account}
\alias{update_account}
\title{Update account}
\usage{
update_account(account_id, body)
}
\arguments{
\item{account_id}{Account Id}

\item{body}{a named list of parameters}
}
\description{
Modifies user’s account.
}
\examples{
\dontrun{
resp <- update_account("123456", list(name = "New account name"))
}
}
\references{
\url{https://developers.coinbase.com/api/v2#update-account}
}
\seealso{
Other wallet-accounts: \code{\link{create_account}},
  \code{\link{delete_account}}, \code{\link{get_accounts}},
  \code{\link{get_account}},
  \code{\link{get_address_transactions}},
  \code{\link{get_address}},
  \code{\link{set_primary_account}}
}
\keyword{internal}
