% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet-users.r
\name{update_current_user}
\alias{update_current_user}
\title{Update current user}
\usage{
update_current_user(body)
}
\arguments{
\item{body}{a named list of parameters}
}
\description{
Modify current user and their preferences.
}
\examples{
\dontrun{
body <- list(name = "mynewname")
resp <- update_current_user(body)
}
}
\references{
\url{https://developers.coinbase.com/api/v2#update-current-user}
}
\seealso{
Other wallet-users: \code{\link{get_auth_info}},
  \code{\link{get_current_user}}, \code{\link{get_user}}
}
\keyword{internal}
