% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{pretty_json}
\alias{pretty_json}
\title{Write json pretty}
\usage{
pretty_json(path)
}
\arguments{
\item{path}{path to json file on disk}
}
\description{
Re-write existing json file with pretty formatting
}
\examples{
\dontrun{
pretty_json("tests/testthat/api.coinbase.com/v2/accounts.json")
}
}
\keyword{internal}
